oracle-apex-mcp/mcp-servers/requirements.txt
fastmcp>=0.2.0
oracledb>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
